requirements.txt
streamlit
langchain-google-genai
python-dotenv
Pillow
langchain-core